+++ gamestore/src/main/resources/application.properties
@@ -1,7 +1,7 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/game_store?useSSL=false&serverTimezone=US/Central&createDatabaseIfNotExist=true&allowPublicKeyRetrieval=true
-spring.datasource.username=root
-spring.datasource.password=password
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
+spring.datasource.url: jdbc:mysql://localhost:3306/game_store?useSSL=false&serverTimezone=UTC&createDatabaseIfNotExist=true&allowPublicKeyRetrieval=true
+spring.datasource.username: root
+spring.datasource.password: password
+spring.jpa.hibernate.ddl-auto: update
+
+
 spring.graphql.graphiql.enabled=true
-spring.graphql.graphiql.path=/graphiql
