+++ gamestore/src/test/java/com/company/gamestore/controllers/ConsoleControllerTest.java
@@ -3,10 +3,9 @@ package com.company.gamestore.controllers;
 import com.company.gamestore.exceptions.NotFoundException;
 import com.company.gamestore.models.Console;
 import com.company.gamestore.repositories.ConsoleRepository;
-import com.company.gamestore.service.ServiceLayer;
+import com.company.gamestore.services.ServiceLayer;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
