+++ gamestore/src/test/java/com/company/gamestore/controllers/InvoiceControllerTest.java
@@ -6,6 +6,7 @@ import com.company.gamestore.repositories.GameRepository;
 import com.company.gamestore.repositories.InvoiceRepository;
 import com.company.gamestore.services.ServiceLayer;
 import com.company.gamestore.viewmodels.InvoiceViewModel;
+import com.company.gamestore.viewmodels.TShirtViewModel;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -14,12 +15,14 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
@@ -137,5 +140,19 @@ public class InvoiceControllerTest {
         mockMvc.perform(get("/invoice/name/{name}","Carl"))
                 .andDo(print())
                 .andExpect(status().isOk());
+    }
+
+    @Test
+    public void shouldReturn422WhenAddingTShirtFails() throws Exception {
+        InvoiceViewModel invoiceViewModel1 = new InvoiceViewModel();
+
+        invoiceViewModel1.setName(null);
+        when(serviceLayer.saveTShirt(any(TShirtViewModel.class))).thenReturn(null);
+
+        mockMvc.perform(MockMvcRequestBuilders
+                        .post("/invoice")
+                        .content(mapper.writeValueAsString(invoiceViewModel))
+                        .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().isUnprocessableEntity());
     }
 }
